moved create AqSession to its own module
